+++ actix-server/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased - 2021-xx-xx
 * Server shutdown would notify all workers to exit regardless if shutdown is graceful.
-  This would make all worker shutdown immediately [#333]
+  This would make all worker shutdown immediately in force shutdown case. [#333]
   
 [#333]: https://github.com/actix/actix-net/pull/333